Refactorizando la estructura del proyecto - Se implementa la estructura de archivos propuesta por el entrenador
