+++ src/app/modules/heroes/components/hero-detail/hero-detail.component.spec.ts
