+++ src/app/modules/heroes/components/hero-detail/hero-detail.component.ts
@@ -2,8 +2,9 @@ import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute } from '@angular/router';
 import { Location } from '@angular/common';
 
-import { Hero } from '../hero';
-import { HeroService } from '../hero.service';
+
+import { Hero } from '../../interfaces/hero';
+import { HeroService } from '../../services/hero/hero.service';
 
 @Component({
   selector: 'app-hero-detail',
