+++ src/app/modules/heroes/components/hero-search/hero-search.component.ts
@@ -5,9 +5,11 @@ import { Observable, Subject } from 'rxjs';
 import {
    debounceTime, distinctUntilChanged, switchMap
  } from 'rxjs/operators';
+import { Hero } from '../../interfaces/hero';
+import { HeroService } from '../../services/hero/hero.service';
 
-import { Hero } from '../hero';
-import { HeroService } from '../hero.service';
+//import { Hero } from '../hero';
+//import { HeroService } from '../hero.service';
 
 @Component({
   selector: 'app-hero-search',
