+++ src/app/modules/heroes/components/messages/messages.component.ts
@@ -1,5 +1,6 @@
 import { Component } from '@angular/core';
-import { MessageService } from '../message.service';
+import { MessageService } from '../../services/message/message.service';
+//import { MessageService } from '../message.service';
 
 @Component({
   selector: 'app-messages',
