+++ src/app/modules/heroes/heroes/heroes.component.ts
@@ -1,7 +1,9 @@
 import { Component, OnInit } from '@angular/core';
+import { Hero } from '../interfaces/hero';
+import { HeroService } from '../services/hero/hero.service';
 
-import { Hero } from '../hero';
-import { HeroService } from '../hero.service';
+//import { Hero } from '../../interfaces/hero';
+//import { HeroService } from '../../services/hero/hero.service';
 
 @Component({
   selector: 'app-heroes',
@@ -35,7 +37,4 @@ export class HeroesComponent implements OnInit {
     this.heroes = this.heroes.filter(h => h !== hero);
     this.heroService.deleteHero(hero.id).subscribe();
   }
-
-
-
 }
