+++ src/app/modules/heroes/interfaces/hero.ts
