+++ src/app/modules/heroes/services/hero/hero.service.ts
@@ -2,9 +2,9 @@ import { Injectable } from '@angular/core';
 
 import { Observable, of } from 'rxjs';
 
-import { Hero } from './hero';
-import { HEROES } from './mock-heroes';
-import { MessageService } from './message.service';
+import { Hero } from '../../interfaces/hero';
+import { HEROES } from '../../mocks/mock-heroes';
+import { MessageService } from '../message/message.service';
 
 import { HttpClient, HttpHeaders } from '@angular/common/http';
