+++ src/app/modules/main/pages/app.component.spec.ts
