+++ src/app/modules/main/pages/app.component.ts
